% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlgorithmStorage.R
\docType{data}
\name{algorithmStorage}
\alias{algorithmStorage}
\title{R6 class to create algorithmStorage objects}
\format{\code{\link{R6Class}} object}
\usage{
algorithmStorage
}
\value{
Algorithm object
}
\description{
An \code{\link[R6]{R6Class}} to save the results from each algorithm object, automatically created 
when a new algorithm object is created.
}
\section{Fields}{

\describe{
\item{\code{config}}{[\code{string}]\cr
An algorithm configuration}

\item{\code{budget}}{[\code{string}]\cr
The current budget}

\item{\code{model}}{[\code{string}]\cr
A model}

\item{\code{performance}}{[\code{string}]\cr
A performance function}
}}

\section{Methods}{

\code{$attachLines(newline)} rbind a new rows \code{newline} to the data.matrix of the algorithmStorage object \cr
}

\keyword{datasets}
