% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BracketStorage.R
\docType{data}
\name{bracketStorage}
\alias{bracketStorage}
\title{R6 class to create bracketStorage objects}
\format{\code{\link{R6Class}} object}
\usage{
bracketStorage
}
\value{
Algorithm object
}
\description{
An \code{\link[R6]{R6Class}} to save the results from each bracket object, automatically created 
when a new bracket object is created.
}
\section{Fields}{

\describe{
\item{\code{models}}{[\code{string}]\cr
A list of models}
}}

\section{Methods}{

\code{$attachLines(newline)} rbinds rows \code{newline} to the data.matrix of the bracketStorage object \cr
}

\keyword{datasets}
