% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperStorage.R
\docType{data}
\name{hyperStorage}
\alias{hyperStorage}
\title{R6 class to create bracketStorage objects}
\format{\code{\link{R6Class}} object}
\usage{
hyperStorage
}
\value{
Algorithm object
}
\description{
An \code{\link[R6]{R6Class}} to save the results from each bracket
}
\section{Fields}{

\describe{
\item{\code{configSpace}}{[\code{string}]\cr
A configuration space constructed with makeParamSet from package ParamHelpers}
}}

\section{Methods}{

\code{$writeDataBase(newline)} rbinds a new row \code{newline} with configurations to the data.matrix  \cr
}

\keyword{datasets}
