% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bracket.R
\docType{data}
\name{bracket}
\alias{bracket}
\title{R6 class to create bracket objects}
\format{\code{\link{R6Class}} object}
\usage{
bracket
}
\value{
[\code{\link{TaskClassif}}].
}
\description{
A \code{\link[R6]{R6Class}} that consits of multiple algorithm objects
}
\section{Fields}{

\describe{
\item{\code{positive}}{[\code{character(1)}]\cr
Only for binary classification: Level of the positive class (\code{NA} otherwise).}

\item{\code{classes}}{[\code{character()}]\cr
Levels of class labels.}

\item{\code{nclasses}}{[\code{integer(1)}]\cr
Number of levels of class labels.}
}}

\examples{
task = TaskClassif$new("iris", data = iris, target = "Species")
task$formula
}
\keyword{datasets}
