% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HyperStorage.R
\docType{data}
\name{hyperStorage}
\alias{hyperStorage}
\title{R6 class to create hyperStorage objects}
\format{\code{\link{R6Class}} object}
\usage{
hyperStorage
}
\value{
Algorithm object
}
\description{
An \code{\link[R6]{R6Class}} to concatenate the results from each bracket object, automatically created 
when the hyperband algorithm is called. Can be used to improve the configuration sample mechanism (e.g. MBO).
Check the vignette for a detailed exampled.
}
\section{Fields}{

\describe{
\item{\code{par.set}}{[\code{string}]\cr
The config space}
}}

\section{Methods}{

\code{$attachLines(newline)} rbinds rows \code{newline} to the data.matrix  \cr
}

\keyword{datasets}
